Define form spec for cells and most fields
Just plugins left now

+++ specifyweb/frontend/js_src/lib/components/ExportFeed/spec.ts
@@ -44,7 +44,7 @@ const itemSpec = f.store(() =>
     metadata: syncers.xmlAttribute('metadata', 'required'),
     days: pipe(
       syncers.xmlAttribute('days', 'required'),
-      syncers.maybe(syncers.toDecimal)
+      syncers.maybe(syncers.toFloat)
     ),
     fileName: syncers.xmlAttribute('filename', 'required'),
     publish: pipe(
